test: RacingCarTest 에 구현한 Name 테스트를 NameTest에 구현
