+++ src/test/java/racingcar/NameTest.java
@@ -1,0 +1,22 @@
+package racingcar;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class NameTest {
+
+    @Test
+    @DisplayName("자동차의 이름은 5자 이하여야 한다.")
+    void name_should_be_less_than_or_equals_five() {
+        // given
+        String name = "greatThanFive";
+
+        // when then
+        assertThatThrownBy(() -> new Name(name))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("자동차의 이름은 5자 이하여야 합니다.");
+    }
+
+}
