+++ src/test/java/racingcar/RacingCarTest.java
@@ -2,12 +2,10 @@ package racingcar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class RacingCarTest {
 
@@ -20,7 +18,7 @@ class RacingCarTest {
 
     @ParameterizedTest
     @ValueSource(ints = {4, 5, 6, 7, 8, 9})
-    @DisplayName("난수 4가 입력되었을 경우 자동차는 한칸 전진한다.")
+    @DisplayName("난수 4이상이 입력되었을 경우 자동차는 한칸 전진한다.")
     void move(int random) {
         // when
         racingCar.moveForward(random);
@@ -31,8 +29,8 @@ class RacingCarTest {
     }
 
     @ParameterizedTest
-    @ValueSource(ints = {1, 2, 3})
-    @DisplayName("난수 3이 입력되었을 경우 자동차는 이동하지 않는다.")
+    @ValueSource(ints = {0, 1, 2, 3})
+    @DisplayName("난수 3이하가 입력되었을 경우 자동차는 이동하지 않는다.")
     void stop(int random) {
         // when
         racingCar.moveForward(random);
@@ -40,17 +38,5 @@ class RacingCarTest {
 
         // then
         assertThat(travelDistance).isZero();
-    }
-
-    @Test
-    @DisplayName("자동차의 이름은 5자 이하여야 한다.")
-    void name_should_be_less_than_or_equals_five() {
-        // given
-        String name = "greatThanFive";
-
-        // when then
-        assertThatThrownBy(() -> new Name(name))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("자동차의 이름은 5자 이하여야 합니다.");
     }
 }
